Legends directory and relocation
