+++ src/components/Legends/RewardLegend/RewardLegend.js
