+++ src/components/Legends/RewardLegend/index.js
